move TelegramLogin to normal components instead of islands advances Preview Bot Login Widget fails to show up 80% of reloads #26
